add pegvm_core
