+++ pegvm/pegvm.c
@@ -15,17 +15,6 @@ uint64_t timer() {
   return tv.tv_sec * 1000 + tv.tv_usec / 1000;
 }
 
-void nez_ShowUsage(const char *file) {
-  // fprintf(stderr, "Usage: %s -f peg_bytecode target_file\n", file);
-  fprintf(stderr, "\npegvm <command> optional files\n");
-  fprintf(stderr, "  -p <filename> Specify an PEGs grammar bytecode file\n");
-  fprintf(stderr, "  -i <filename> Specify an input file\n");
-  fprintf(stderr, "  -o <filename> Specify an output file\n");
-  fprintf(stderr, "  -t <type>     Specify an output type\n");
-  fprintf(stderr, "  -h            Display this help and exit\n\n");
-  exit(EXIT_FAILURE);
-}
-
 void nez_PrintErrorInfo(const char *errmsg) {
   fprintf(stderr, "%s\n", errmsg);
   exit(EXIT_FAILURE);
@@ -300,157 +289,6 @@ void nez_DisposeParsingContext(ParsingContext ctx) {
   // dispose_pego(&ctx->unusedObject);
 }
 
-// static inline int ParserContext_IsFailure(ParsingContext context)
-//{
-//    return context->left == NULL;
-//}
-//
-// static void ParserContext_RecordFailurePos(ParsingContext context, size_t
-// consumed)
-//{
-//    context->left = NULL;
-//    context->pos -= consumed;
-//}
-
-// #define INC_SP(N) (context->stack_pointer += (N))
-// #define DEC_SP(N) (context->stack_pointer -= (N))
-static inline long INC_SP(ParsingContext context, int N) {
-  context->stack_pointer += (N);
-#if PEGVM_DEBUG
-  assert(context->stack_pointer >= context->stack_pointer_base &&
-         context->stack_pointer <
-             &context->stack_pointer_base[PARSING_CONTEXT_MAX_STACK_LENGTH]);
-#endif
-  return *context->stack_pointer;
-}
-
-static inline long DEC_SP(ParsingContext context, int N) {
-  context->stack_pointer -= N;
-#if PEGVM_DEBUG
-  assert(context->stack_pointer >= context->stack_pointer_base &&
-         context->stack_pointer <
-             &context->stack_pointer_base[PARSING_CONTEXT_MAX_STACK_LENGTH]);
-#endif
-  return *context->stack_pointer;
-}
-
-static inline ParsingObject INC_OSP(ParsingContext context, int N) {
-  context->object_stack_pointer += (N);
-#if PEGVM_DEBUG
-  assert(context->object_stack_pointer >= context->object_stack_pointer_base &&
-         context->object_stack_pointer <
-             &context->object_stack_pointer_base
-                  [PARSING_CONTEXT_MAX_STACK_LENGTH]);
-#endif
-  return *context->object_stack_pointer;
-}
-
-static inline ParsingObject DEC_OSP(ParsingContext context, int N) {
-  context->object_stack_pointer -= N;
-#if PEGVM_DEBUG
-  assert(context->object_stack_pointer >= context->object_stack_pointer_base &&
-         context->object_stack_pointer <
-             &context->object_stack_pointer_base
-                  [PARSING_CONTEXT_MAX_STACK_LENGTH]);
-#endif
-  return *context->object_stack_pointer;
-}
-
-static inline void PUSH_IP(ParsingContext context, Instruction *pc) {
-  *context->call_stack_pointer++ = pc;
-#if PEGVM_DEBUG
-  assert(
-      context->call_stack_pointer >= context->call_stack_pointer_base &&
-      context->call_stack_pointer <
-          &context->call_stack_pointer_base[PARSING_CONTEXT_MAX_STACK_LENGTH]);
-#endif
-}
-
-static inline Instruction **POP_IP(ParsingContext context) {
-  --context->call_stack_pointer;
-  assert(
-      context->call_stack_pointer >= context->call_stack_pointer_base &&
-      context->call_stack_pointer <
-          &context->call_stack_pointer_base[PARSING_CONTEXT_MAX_STACK_LENGTH]);
-  return context->call_stack_pointer;
-}
-
-#define SP_TOP(INST) (*context->stack_pointer)
-#define PUSH_SP(INST) (*context->stack_pointer = (INST), INC_SP(context, 1))
-#define POP_SP(INST) (DEC_SP(context, 1))
-#define PUSH_OSP(INST) \
-  (*context->object_stack_pointer = (INST), INC_OSP(context, 1))
-#define POP_OSP(INST) (DEC_OSP(context, 1))
-//#define TOP_SP() (*context->stack_pointer)
-#define JUMP       \
-  pc = (pc)->jump; \
-  goto *(pc)->ptr;
-#define RET              \
-  pc = *POP_IP(context); \
-  goto *(pc)->ptr;
-
-#if PEGVM_PROFILE
-static uint64_t count[PEGVM_OP_MAX];
-static uint64_t conbination_count[PEGVM_OP_MAX][PEGVM_OP_MAX];
-static uint64_t count_all;
-static uint64_t rule_count[100];
-#define DISPATCH_NEXT                          \
-  int first = (int)pc->opcode;                 \
-  ++pc;                                        \
-  conbination_count[first][(int)pc->opcode]++; \
-  goto *(pc)->ptr;
-#define OP(OP)                            \
-  PEGVM_OP_##OP : count[PEGVM_OP_##OP]++; \
-  count_all++;
-#else
-#define OP(OP) PEGVM_OP_##OP:
-#define DISPATCH_NEXT \
-  ++pc;               \
-  goto *(pc)->ptr;
-#endif
-
-static const char *get_json_rule(uint8_t json_rule) {
-  switch (json_rule) {
-#define json_CASE(RULE)           \
-  case PEGVM_PROFILE_json_##RULE: \
-    return "" #RULE;
-    PEGVM_PROFILE_json_EACH(json_CASE);
-  default:
-    assert(0 && "UNREACHABLE");
-    break;
-#undef json_CASE
-  }
-  return "";
-}
-
-static const char *get_xml_rule(uint8_t xml_rule) {
-  switch (xml_rule) {
-#define xml_CASE(RULE)           \
-  case PEGVM_PROFILE_xml_##RULE: \
-    return "" #RULE;
-    PEGVM_PROFILE_xml_EACH(xml_CASE);
-  default:
-    assert(0 && "UNREACHABLE");
-    break;
-#undef xml_CASE
-  }
-  return "";
-}
-
-static const char *get_c99_rule(uint8_t c99_rule) {
-  switch (c99_rule) {
-#define c99_CASE(RULE)           \
-  case PEGVM_PROFILE_c99_##RULE: \
-    return "" #RULE;
-    PEGVM_PROFILE_c99_EACH(c99_CASE);
-  default:
-    assert(0 && "UNREACHABLE");
-    break;
-#undef c99_CASE
-  }
-  return "";
-}
-
 void nez_VM_PrintProfile(const char *file_type) {
 #if PEGVM_PROFILE
   fprintf(stderr, "\ninstruction count \n");
@@ -482,6 +320,7 @@ void nez_VM_PrintProfile(const char *file_type) {
     fprintf(file, "\n");
   }
   fclose(file);
+#if 0
   if (file_type) {
     fprintf(stderr, "\nrule_count\n");
     if (!strcmp(file_type, "json")) {
@@ -499,7 +338,10 @@ void nez_VM_PrintProfile(const char *file_type) {
     }
   }
 #endif
+#endif
 }
+
+#include "pegvm_core.c"
 
 ParsingObject nez_Parse(ParsingContext context, Instruction *inst) {
   if (nez_VM_Execute(context, inst)) {
@@ -539,412 +381,11 @@ void nez_Match(ParsingContext context, Instruction *inst) {
 Instruction *nez_VM_Prepare(ParsingContext context, Instruction *inst) {
   long i;
   const void **table = (const void **)nez_VM_Execute(context, NULL);
+  Instruction *ip = inst;
   for (i = 0; i < context->bytecode_length; i++) {
-    (inst + i)->ptr = table[(inst + i)->opcode];
+    ip->ptr = table[ip->opcode];
+    ++ip;
   }
   return inst;
 }
 
-long nez_VM_Execute(ParsingContext context, Instruction *inst) {
-  static const void *table[] = {
-#define DEFINE_TABLE(NAME) &&PEGVM_OP_##NAME,
-    PEGVM_OP_EACH(DEFINE_TABLE)
-#undef DEFINE_TABLE
-  };
-
-  if (inst == NULL) {
-    return (long)table;
-  }
-
-  int failflag = 0;
-  ParsingObject left = context->left;
-  long pos = context->pos;
-  Instruction *pc = inst + context->startPoint;
-  const char *inputs = context->inputs;
-  MemoryPool pool = context->mpool;
-
-  PUSH_IP(context, inst);
-  P4D_setObject(context, &left, P4D_newObject(context, context->pos, pool));
-
-  long Reg1 = 0;
-  long Reg2 = 0;
-  long Reg3 = 0;
-
-  goto *(pc)->ptr;
-
-  OP(EXIT) {
-    P4D_commitLog(context, 0, left, pool);
-    context->left = left;
-    context->pos = pos;
-    return failflag;
-  }
-  OP(JUMP) { JUMP; }
-  OP(CALL) {
-    PUSH_IP(context, pc + 1);
-#if PEGVM_PROFILE
-    rule_count[pc->ndata[0]]++;
-#endif
-    JUMP;
-    // pc = inst[pc].jump;
-    // goto *inst[pc].ptr;
-  }
-  OP(RET) { RET; }
-  OP(CONDBRANCH) {
-    if (failflag == *pc->ndata) {
-      JUMP;
-    }
-    DISPATCH_NEXT;
-  }
-  OP(REPCOND) {
-    if (pos != POP_SP()) {
-      DISPATCH_NEXT;
-    }
-    JUMP;
-  }
-  OP(CHARRANGE) {
-    if ((pc)->chardata[0] <= inputs[pos] && inputs[pos] <= (pc)->chardata[1]) {
-      pos++;
-      DISPATCH_NEXT;
-    } else {
-      failflag = 1;
-      JUMP;
-    }
-  }
-  OP(CHARSET) {
-    int j = 0;
-    int len = *(pc)->ndata;
-    while (j < len) {
-      if (inputs[pos] == (pc)->chardata[j]) {
-        pos++;
-        DISPATCH_NEXT;
-      }
-      j++;
-    }
-    failflag = 1;
-    JUMP;
-  }
-  OP(STRING) {
-    char *p = pc->chardata;
-    int len = *pc->ndata;
-    char *pend = pc->chardata + len;
-    while (p < pend) {
-      if (inputs[pos] != *p++) {
-        failflag = 1;
-        JUMP;
-      }
-      pos++;
-    }
-    DISPATCH_NEXT;
-  }
-  OP(ANY) {
-    if (inputs[pos++] == 0) {
-      pos--;
-      failflag = 1;
-      JUMP;
-    }
-    DISPATCH_NEXT;
-  }
-  OP(PUSHo) {
-    ParsingObject po = P4D_newObject(context, pos, pool);
-    *po = *left;
-    po->refc = 1;
-    PUSH_OSP(po);
-    DISPATCH_NEXT;
-  }
-  OP(PUSHconnect) {
-    ParsingObject po = left;
-    left->refc++;
-    PUSH_OSP(po);
-    PUSH_SP(P4D_markLogStack(context));
-    DISPATCH_NEXT;
-  }
-  OP(PUSHp1) {
-    PUSH_SP(pos);
-    DISPATCH_NEXT;
-  }
-  OP(LOADp1) {
-    Reg1 = pos;
-    DISPATCH_NEXT;
-  }
-  OP(LOADp2) {
-    Reg2 = pos;
-    DISPATCH_NEXT;
-  }
-  OP(LOADp3) {
-    Reg3 = pos;
-    DISPATCH_NEXT;
-  }
-  OP(POPp) {
-    POP_SP();
-    DISPATCH_NEXT;
-  }
-  OP(POPo) {
-    ParsingObject po = POP_OSP();
-    P4D_setObject(context, &po, NULL);
-    DISPATCH_NEXT;
-  }
-  OP(STOREo) {
-    ParsingObject po = POP_OSP();
-    P4D_setObject(context, &left, po);
-    DISPATCH_NEXT;
-  }
-  OP(STOREp) {
-    pos = POP_SP();
-    DISPATCH_NEXT;
-  }
-  OP(STOREp1) {
-    pos = Reg1;
-    DISPATCH_NEXT;
-  }
-  OP(STOREp2) {
-    pos = Reg2;
-    DISPATCH_NEXT;
-  }
-  OP(STOREp3) {
-    pos = Reg3;
-    DISPATCH_NEXT;
-  }
-  OP(STOREflag) {
-    failflag = pc->ndata[0];
-    DISPATCH_NEXT;
-  }
-  OP(NEW) {
-    PUSH_SP(P4D_markLogStack(context));
-    P4D_setObject(context, &left, P4D_newObject(context, pos, pool));
-    // PUSH_SP(P4D_markLogStack(context));
-    DISPATCH_NEXT;
-  }
-  OP(NEWJOIN) {
-    ParsingObject po = NULL;
-    P4D_setObject(context, &po, left);
-    P4D_setObject(context, &left, P4D_newObject(context, pos, pool));
-    // PUSH_SP(P4D_markLogStack(context));
-    P4D_lazyJoin(context, po, pool);
-    P4D_lazyLink(context, left, *(pc)->ndata, po, pool);
-    DISPATCH_NEXT;
-  }
-  OP(COMMIT) {
-    P4D_commitLog(context, (int)POP_SP(), left, pool);
-    ParsingObject parent = (ParsingObject)POP_OSP();
-    P4D_lazyLink(context, parent, *(pc)->ndata, left, pool);
-    P4D_setObject(context, &left, parent);
-    DISPATCH_NEXT;
-  }
-  OP(ABORT) {
-    P4D_abortLog(context, (int)POP_SP());
-    DISPATCH_NEXT;
-  }
-  OP(LINK) {
-    ParsingObject parent = (ParsingObject)POP_OSP();
-    P4D_lazyLink(context, parent, *(pc)->ndata, left, pool);
-    // P4D_setObject(context, &left, parent);
-    PUSH_OSP(parent);
-    DISPATCH_NEXT;
-  }
-  OP(SETendp) {
-    left->end_pos = pos;
-    DISPATCH_NEXT;
-  }
-  OP(TAG) {
-    left->tag = (pc)->chardata;
-    DISPATCH_NEXT;
-  }
-  OP(VALUE) {
-    left->value = (pc)->chardata;
-    DISPATCH_NEXT;
-  }
-  OP(MAPPEDCHOICE) {
-    pc = inst + (pc)->ndata[(int)inputs[pos]];
-    goto *(pc)->ptr;
-  }
-  OP(SCAN) {
-    long start = POP_SP();
-    long len = pos - start;
-    char value[len];
-    int j = 0;
-    for (long i = start; i < pos; i++) {
-      value[j] = inputs[i];
-      j++;
-    }
-    if (pc->ndata[0] == 16) {
-      long num = strtol(value, NULL, 16);
-      context->repeat_table[pc->ndata[1]] = (int)num;
-      DISPATCH_NEXT;
-    }
-    context->repeat_table[pc->ndata[1]] = atoi(value);
-    DISPATCH_NEXT;
-  }
-  OP(CHECKEND) {
-    if (context->repeat_table[pc->ndata[0]] == 0) {
-      DISPATCH_NEXT;
-    }
-    JUMP;
-  }
-  OP(DEF) {
-    long start = POP_SP();
-    int len = (int)(pos - start);
-    char *value = malloc(len);
-    int j = 0;
-    for (long i = start; i < pos; i++) {
-      value[j] = inputs[i];
-      j++;
-    }
-    pushSymbolTable(context, pc->ndata[0], len, value);
-    DISPATCH_NEXT;
-  }
-  OP(IS) {
-    failflag = matchSymbolTableTop(context, &pos, pc->ndata[0]);
-    DISPATCH_NEXT;
-  }
-  OP(ISA) {
-    failflag = matchSymbolTable(context, &pos, pc->ndata[0]);
-    DISPATCH_NEXT;
-  }
-  OP(BLOCKSTART) {
-    long len;
-    PUSH_SP(context->stateValue);
-    char *value = getIndentText(context, inputs, pos, &len);
-    PUSH_SP(pushSymbolTable(context, pc->ndata[0], (int)len, value));
-    DISPATCH_NEXT;
-  }
-  OP(BLOCKEND) {
-    popSymbolTable(context, (int)POP_SP());
-    context->stateValue = (int)POP_SP();
-    DISPATCH_NEXT;
-  }
-  OP(INDENT) {
-    matchSymbolTableTop(context, &pos, pc->ndata[0]);
-    DISPATCH_NEXT;
-  }
-  OP(NOTBYTE) {
-    if (inputs[pos] != *(pc)->ndata) {
-      DISPATCH_NEXT;
-    }
-    failflag = 1;
-    JUMP;
-  }
-  OP(NOTANY) {
-    if (inputs[pos] != 0) {
-      failflag = 1;
-      JUMP;
-    }
-    DISPATCH_NEXT;
-  }
-  OP(NOTCHARSET) {
-    int j = 0;
-    int len = *(pc)->ndata;
-    while (j < len) {
-      if (inputs[pos] == (pc)->chardata[j]) {
-        failflag = 1;
-        JUMP;
-      }
-      j++;
-    }
-    DISPATCH_NEXT;
-  }
-  OP(NOTBYTERANGE) {
-    if (!(inputs[pos] >= (pc)->chardata[0] &&
-          inputs[pos] <= (pc)->chardata[1])) {
-      DISPATCH_NEXT;
-    }
-    failflag = 1;
-    JUMP;
-  }
-  OP(NOTSTRING) {
-    char *p = pc->chardata;
-    int len = *pc->ndata;
-    long backtrack_pos = pos;
-    char *pend = pc->chardata + len;
-    while (p < pend) {
-      if (inputs[pos] != *p++) {
-        pos = backtrack_pos;
-        DISPATCH_NEXT;
-      }
-      pos++;
-    }
-    pos = backtrack_pos;
-    failflag = 1;
-    JUMP;
-  }
-  OP(OPTIONALBYTE) {
-    if (inputs[pos] == *(pc)->ndata) {
-      pos++;
-    }
-    DISPATCH_NEXT;
-  }
-  OP(OPTIONALCHARSET) {
-    int j = 0;
-    int len = *(pc)->ndata;
-    while (j < len) {
-      if (inputs[pos] == (pc)->chardata[j]) {
-        pos++;
-        DISPATCH_NEXT;
-      }
-      j++;
-    }
-    DISPATCH_NEXT;
-  }
-  OP(OPTIONALBYTERANGE) {
-    if (inputs[pos] >= (pc)->chardata[0] && inputs[pos] <= (pc)->chardata[1]) {
-      pos++;
-    }
-    DISPATCH_NEXT;
-  }
-  OP(OPTIONALSTRING) {
-    char *p = pc->chardata;
-    int len = *pc->ndata;
-    long backtrack_pos = pos;
-    char *pend = pc->chardata + len;
-    while (p < pend) {
-      if (inputs[pos] != *p++) {
-        pos = backtrack_pos;
-        DISPATCH_NEXT;
-      }
-      pos++;
-    }
-    DISPATCH_NEXT;
-  }
-  OP(ZEROMOREBYTERANGE) {
-    while (1) {
-      if (!(inputs[pos] >= (pc)->chardata[0] &&
-            inputs[pos] <= (pc)->chardata[1])) {
-        DISPATCH_NEXT;
-      }
-      pos++;
-    }
-  }
-  OP(ZEROMORECHARSET) {
-    int j = 0;
-    int len = *(pc)->ndata;
-    while (j < len) {
-      if (inputs[pos] == (pc)->chardata[j]) {
-        pos++;
-        j = 0;
-      } else {
-        j++;
-      }
-    }
-    DISPATCH_NEXT;
-  }
-  OP(ZEROMOREWS) {
-    while (1) {
-      char c = inputs[pos];
-      if (c == 32 || c == 9 || c == 10 || c == 13) {
-        pos++;
-      } else {
-        break;
-      }
-    }
-    DISPATCH_NEXT;
-  }
-  OP(REPEATANY) {
-    long back = pos;
-    pos = pos + context->repeat_table[pc->ndata[0]];
-    if (pos - 1 > (long)context->input_size) {
-      pos = back;
-    }
-    DISPATCH_NEXT;
-  }
-
-  return failflag;
-}
